move to postgress

+++ DBmodels/DBUser.py
@@ -1,6 +1,5 @@
-from sqlalchemy import Column, Integer, String, ARRAY, Table, ForeignKey
+from sqlalchemy import Column, Integer, String, BigInteger
 from sqlalchemy.orm import relationship
-from DBmodels.association import user_bookclub
 
 from database import Base
 
@@ -9,7 +8,7 @@ class DBUser(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String, nullable=False)
-    phone_number = Column(Integer, unique=True, nullable=False)
+    phone_number = Column(BigInteger, unique=True, nullable=False)
     password = Column(String, nullable=False)
     access_token = Column(String, unique=True, nullable=False)
     book_clubs = relationship("DBBookClub", secondary="user_bookclub", back_populates="members")
